requirements.txt
openai==1.36.1
python-dotenv==1.0.1
jsonlines==4.0.0
rouge-score==0.1.2
jieba==0.42.1